Protocol change: 16 bit pressure value
This is for future proofing. Internally, only 256 pressure levels are used at the moment.

+++ src/shared/net/pen.h
@@ -92,9 +92,9 @@ private:
 
 struct PenPoint {
 	PenPoint() {}
-	PenPoint(int32_t x_, int32_t y_, uint8_t p_) : x(x_), y(y_), p(p_) {}
+	PenPoint(int32_t x_, int32_t y_, uint16_t p_) : x(x_), y(y_), p(p_) {}
 	int32_t x, y;
-	uint8_t p;
+	uint16_t p;
 };
 
 typedef QVector<PenPoint> PenPointVector;
